Add polcal option
This adds the --polcal option and removes the --XYfix option. So far --polcal
will cause the pipeline to search for the delay calibration associated with the
subarray of the observation and run MKXPhase to get the bandpass phases on the
Y polarisation. These bandpass phases are then merged with the bandpass table
into a new bptable using BPCal and this new BP table is applied to all
subsequent data.